+++ tests/test_manifest.js
@@ -58,6 +58,7 @@ export const schemeTestFiles = [
     'tests/core/scheme/exception_tests.scm',
     'tests/core/scheme/error_tests.scm',
     'tests/core/scheme/repl_tests.scm',
+    'tests/core/scheme/parameter_tests.scm',
 ];
 
 /**